+++ product/entityes.go
@@ -4,8 +4,8 @@ import "time"
 
 type Product struct {
 	ProductName   string    `json:"product_name"`
-	Weight        int       `json:"weight"`
-	Calories      int       `json:"calories"`
+	Weight        float64   `json:"weight"`
+	Calories      float64   `json:"calories"`
 	Proteins      float64   `json:"proteins"`
 	Fats          float64   `json:"fats"`
 	Carbohydrates float64   `json:"carbohydrates"`
